+++ src/main/java/cart/controller/support/BasicAuth.java
@@ -1,4 +1,4 @@
-package cart.advice;
+package cart.controller.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
