+++ src/main/java/cart/controller/support/BasicAuthResolver.java
@@ -1,4 +1,4 @@
-package cart.advice;
+package cart.controller.support;
 
 import static org.apache.tomcat.util.codec.binary.Base64.decodeBase64;
