+++ src/main/java/cart/controller/support/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package cart.advice;
+package cart.controller.support;
 
 import cart.dto.ApiErrorResponse;
 import cart.exception.InvalidPasswordException;
